Separate map and app reducers
