+++ src/model/app/reducers.js
@@ -1,0 +1,10 @@
+const initialState = {};
+
+export default function app(state = initialState, action) {
+  switch (action.type) {
+    default:
+      return {
+        ...state,
+      };
+  }
+}
