+++ src/model/map/actions.js
@@ -1,8 +1,8 @@
-export const SET_LAYERS = 'SET_LAYERS';
+export const ADD_LAYERS = 'ADD_LAYERS';
 export const SET_CENTER = 'SET_CENTER';
+export const SET_LAYERS = 'SET_LAYERS';
 export const SET_RESOLUTION = 'SET_RESOLUTION';
 export const SET_ZOOM = 'SET_ZOOM';
-export const ADD_LAYERS = 'ADD_LAYERS';
 
 export const setLayers = data => ({ type: SET_LAYERS, data });
