+++ src/model/map/reducers.js
